LayoutTests/imported/w3c/web-platform-tests/intersection-observer/observer-without-js-reference-expected.txt


PASS IntersectionObserver that is unreachable in js should still generate notifications. 
PASS First rAF 
FAIL document.scrollingElement.scrollTop = 300 assert_equals: Two notifications. expected 2 but got 1
